Extract SharedElement to its own file
