+++ examples/NavigationPlayground/js/Transitions/SharedElements/index.js
@@ -8,59 +8,7 @@ import { Transition } from 'react-navigation';
 import _ from 'lodash';
 import faker from 'faker';
 
-const {createTransition, together} = Transition;
-
-const SharedElements = (filter) => ({
-  filter,
-  getItemsToClone(
-    itemsOnFromRoute: Array<*>, 
-    itemsOnToRoute: Array<*> ) {
-    const itemIdsOnBoth = _.intersectionWith(itemsOnFromRoute, itemsOnToRoute, (i1, i2) => i1.id === i2.id)
-      .map(item => item.id);
-    const onBoth = item => itemIdsOnBoth.includes(item.id);
-    return itemsOnFromRoute.filter(onBoth);
-  },
-  getItemsToMeasure(
-    itemsOnFromRoute: Array<*>, 
-    itemsOnToRoute: Array<*> ) {
-    const itemIdsOnBoth = _.intersectionWith(itemsOnFromRoute, itemsOnToRoute, (i1, i2) => i1.id === i2.id)
-      .map(item => item.id);
-    const onBoth = item => itemIdsOnBoth.includes(item.id);
-    return itemsOnFromRoute.filter(onBoth).concat(itemsOnToRoute.filter(onBoth));
-  },
-  createAnimatedStyleMapForClones(
-    itemsOnFromRoute: Array<*>, 
-    itemsOnToRoute: Array<*>, 
-    transitionProps) {
-    const itemIdsOnBoth = _.intersectionWith(itemsOnFromRoute, itemsOnToRoute, (i1, i2) => i1.id === i2.id)
-      .map(item => item.id);
-    const {progress} = transitionProps;
-    const createSharedItemStyle = (result, id) => {
-      const fromItem = itemsOnFromRoute.find(item => item.id === id);
-      const toItem = itemsOnToRoute.find(item => item.id === id);
-      if (!fromItem.isMeasured() || !toItem.isMeasured()) return null; // TODO temporary line to avoid crashes, should remove
-      console.log('fromItem', fromItem.toString(), 'toItem', toItem.toString());
-      const inputRange = [0, 1];
-      const left = progress.interpolate({
-        inputRange, outputRange: [fromItem.metrics.x, toItem.metrics.x]
-      });
-      const top = progress.interpolate({
-        inputRange, outputRange: [fromItem.metrics.y, toItem.metrics.y]
-      });
-      const width = progress.interpolate({
-        inputRange, outputRange: [fromItem.metrics.width, toItem.metrics.width]
-      });
-      const height = progress.interpolate({
-        inputRange, outputRange: [fromItem.metrics.height, toItem.metrics.height]
-      });
-      result[id] = { left, top, width, height, right: null, bottom: null };
-      return result;
-    };
-    return {
-      from: itemIdsOnBoth.reduce(createSharedItemStyle, {}),
-    }
-  }
-});
+const {createTransition, together, Transitions} = Transition;
 
 const CrossFade = (filter) => ({
   filter,
@@ -129,7 +77,7 @@ const FastFadeOutFromRoute = (filter) => ({
   }
 });
 
-const SharedImage = createTransition(SharedElements, /image-.+/);
+const SharedImage = createTransition(Transitions.SharedElement, /image-.+/);
 const CrossFadeScene = createTransition(CrossFade, /\$scene.+/);
 
 const DelayedFadeInDetail = createTransition(DelayedFadeInToRoute, /\$scene-PhotoDetail/);
@@ -202,10 +150,10 @@ const StaggeredAppearImages = createTransition(StaggeredAppear, /image-.+/);
 const transitions = [
   // { from: 'PhotoGrid', to: 'PhotoDetail', transition: CrossFadeScene },
   // { from: 'PhotoDetail', to: 'PhotoGrid', transition: CrossFadeScene },
-  // { from: 'PhotoGrid', to: 'PhotoDetail', transition: together(SharedImage, DelayedFadeInDetail)},
-  // { from: 'PhotoDetail', to: 'PhotoGrid', transition: together(SharedImage, FastFadeOutDetail) },
-  { from: 'PhotoGrid', to: 'PhotoDetail', transition: CrossFadeScene },
-  { from: 'PhotoDetail', to: 'PhotoGrid', transition: together(StaggeredAppearImages, SlideScenes) },
+  { from: 'PhotoGrid', to: 'PhotoDetail', transition: together(SharedImage, DelayedFadeInDetail)},
+  { from: 'PhotoDetail', to: 'PhotoGrid', transition: together(SharedImage, FastFadeOutDetail) },
+  // { from: 'PhotoGrid', to: 'PhotoDetail', transition: CrossFadeScene },
+  // { from: 'PhotoDetail', to: 'PhotoGrid', transition: together(StaggeredAppearImages, SlideScenes) },
 ];
 
 const App = StackNavigator({
