+++ src/views/Transition/Transitions/index.js
@@ -1,0 +1,5 @@
+import SharedElement from './SharedElement';
+
+export default {
+  SharedElement,
+}
